Retreve Schedules Endpoint

+++ fishPI/services/lighting.py
@@ -1,40 +1,49 @@
 import fishPI
+import os.path
 from fishPI import logging
 from fishPI import services
 from fishPI import models
 
 def load():
-    logging.logInfo(" * Loading lighting service")
-
     channel = 0
     while channel < len(fishPI.config.light_pins) :
         channel = channel + 1
         services.database.set_initial("{}_brightness".format(channel),0)
-        services.database.set_initial("{}_schedule".format(channel),models.light_schedule.light_schedule().toJSON())
+        services.database.set_initial("{}_schedule".format(channel),models.lighting.schedule().as_json())
 
 def light_channel_meta(channel):
-    return "brightness_channel_{}".format(channel)
+    return "{}_brightness".format(channel)
 
+def schedule_channel_meta(channel):
+    return "{}_schedule".format(channel)
+                         
 def channel_to_pin(channel):
-    return fishPI.config.light_pins[channel]
+    return fishPI.config.light_pins[int(channel)]
+
+def get_schedule(channel):
+    return fishPI.services.database.get_meta(schedule_channel_meta(channel))
 
 def get_brightness(channel):
-    return fishPI.services.database.get_meta(light_channel_meta(channel))["value"]
+    return fishPI.services.database.get_meta(light_channel_meta(channel))
 
 def set_brightness(channel, percentage):
-    current_percentage = get_brightness(channel)
+    current_percentage = float(get_brightness(channel).value)
+    percentage = float(percentage)
 
-    if(percentage == current_percentage): return True
+    if(percentage == current_percentage): get_brightness(channel) 
 
-    change_per_increment = current_percentage - percentage / 30
+    if(current_percentage < percentage):
+        change_per_increment = int(current_percentage) + int(percentage) / 30
+    else:
+        change_per_increment = int(current_percentage) - int(percentage) / 30
 
     x = 0
     while x <= 30:
-        x = X + 1
+        x = x + 1
         pi_blaster(channel_to_pin(channel), current_percentage + (change_per_increment * x))
 
-    fishPI.services.database.save_meta(light_channel_meta(channel),percentage)
-    return True
+    return fishPI.services.database.set_meta(light_channel_meta(channel),str(percentage))
+    
 
 def flash_channel(channel):
         pi_blaster(channel_to_pin(channel), 0)
@@ -42,7 +51,17 @@ def flash_channel(channel):
         pi_blaster(channel_to_pin(channel), 0)
 
 def pi_blaster(pin,percentage):
-    v = (percentage / 100)
-    with open('/dev/pi-blaster', 'w') as printer:
-        printer.write("{} = {}".format(pin,v))
 
+    if(percentage > 100):
+        percentage = 100
+
+    if(percentage < 0):
+        percentage = 0
+
+    v = round((percentage / 100),1)
+   
+    if(os.path.exists('/dev/pi-blaster')):
+        with open('/dev/pi-blaster', 'w') as printer:
+            printer.write("{} = {}".format(pin,v))
+    else:
+        print("Blasting Pin {} with {}".format(pin,v))
